# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tagg@d API",
    "version": "1.0.0",
    "description": "API documentation for the Tagg@d backend — a crypto payment platform for Africa."
  },
  "servers": [
    {
      "url": "http://localhost:5002/api/v2",
      "description": "Current version (v2)"
    },
    {
      "url": "http://localhost:5002/api/v1",
      "description": "Deprecated version (v1)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Development Server (unversioned root)"
    },
    {
      "url": "https://taggedpay.xyz/api/v2",
      "description": "Production (v2)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 42
          },
          "user_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "credit",
              "debit",
              "payment",
              "swap"
            ],
            "example": "payment"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed"
            ],
            "example": "completed"
          },
          "amount": {
            "type": "number",
            "example": 50
          },
          "usd_value": {
            "type": "number",
            "example": 50
          },
          "chain": {
            "type": "string",
            "example": "stellar"
          },
          "token": {
            "type": "string",
            "example": "USDC"
          },
          "sender_tag": {
            "type": "string",
            "example": "alice"
          },
          "receiver_tag": {
            "type": "string",
            "example": "bob"
          },
          "tx_hash": {
            "type": "string",
            "example": "0xabc123..."
          },
          "note": {
            "type": "string",
            "example": "Monthly rent"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/tags/search": {
      "get": {
        "summary": "Search for tags",
        "description": "Search registered tags by name. Rate-limited to 10 requests per minute per IP.",
        "tags": [
          "Tags"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query string"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tag": {
                            "type": "string",
                            "example": "alice"
                          },
                          "walletAddress": {
                            "type": "string"
                          },
                          "chain": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/api/transactions/search": {
      "get": {
        "summary": "Search transactions with filters",
        "tags": [
          "Transaction Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Full-text search string"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "completed",
                "pending",
                "failed"
              ]
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "credit",
                "debit",
                "payment",
                "swap"
              ]
            }
          },
          {
            "in": "query",
            "name": "chain",
            "schema": {
              "type": "string",
              "default": 20,
              "enum": [
                "XLM",
                "BASE",
                "LSK",
                "FLOW",
                "U2U",
                "STRK"
              ]
            }
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (ISO format)"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (ISO format)"
          },
          {
            "in": "query",
            "name": "minAmount",
            "schema": {
              "type": "number"
            },
            "description": "Minimum USD value"
          },
          {
            "in": "query",
            "name": "maxAmount",
            "schema": {
              "type": "number"
            },
            "description": "Maximum USD value"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "date",
                "amount",
                "relevance"
              ],
              "default": "relevance"
            }
          },
          {
            "in": "query",
            "name": "sortDir",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor for next page"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "type": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "usd_value": {
                            "type": "number"
                          },
                          "chain": {
                            "type": "string"
                          },
                          "token": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "next_cursor": {
                          "type": "string"
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Search rate limit exceeded (30 per minute)"
          }
        },
        "description": "Full-text and filtered search across transaction history. Supports keyword search, status/chain/token filters, amount ranges, and date ranges. Returns cursor-based paginated results."
      }
    },
    "/api/transactions": {
      "get": {
        "summary": "Get user's transaction history",
        "description": "Returns paginated transaction history for the authenticated user. Supports `transactions:read` scope for API keys.",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get a specific transaction by ID",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Transaction ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "put": {
        "summary": "Update a transaction (e.g., add a note)",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction updated"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a transaction",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction deleted"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
    "/api/transactions/search/export": {
      "get": {
        "summary": "Export search results as CSV",
        "description": "Same filter parameters as search, but returns all matching results as a downloadable CSV file instead of paginated JSON.",
        "tags": [
          "Transaction Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "completed",
                "pending",
                "failed"
              ]
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "credit",
                "debit",
                "payment",
                "swap"
              ]
            }
          },
          {
            "in": "query",
            "name": "chain",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "minAmount",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "maxAmount",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file download",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Search rate limit exceeded"
          }
        }
      }
    },
...
  },
...
}